requirements.txt
plotly==5.24.1
pandas==2.2.3
dash-bootstrap-components
wordcloud==1.9.3